requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
openai>=1.30.0
FlagEmbedding>=1.2.0
# --- 多源 loader 依赖（PDF / DOCX / HTML / URL，按需安装）---
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
